Renamed Client and Model to prevent future naming conflicts.
Since Client and Model are popular class names, it's likely that when using the class, it would need to
be aliased. For example: use Docusign\Client; might need to be use Docusign\Client as DocuSignClient in
order to prevent a potential naming conflict. So to prevent this, I'm leaving the DocuSign prefix.